moves similar documents to peachjam_ml
