+++ lawlibrary/settings.py
@@ -3,6 +3,7 @@ from liiweb.settings import *  # noqa
 INSTALLED_APPS = [
     "lawlibrary.apps.LawlibraryConfig",
     "peachjam_subs",
+    "peachjam_ml",
 ] + INSTALLED_APPS  # noqa
 
 ROOT_URLCONF = "lawlibrary.urls"
