+++ lawlibrary/urls.py
@@ -42,6 +42,7 @@ urlpatterns = [
             ]
         ),
     ),
+    path("", include("peachjam_ml.urls")),
     path("", include("liiweb.urls")),
     # this is duplicated here because it overrides the liiweb url with the same name, and django uses the
     # last occurrence when looking up a url, but the first when dispatching
