+++ peachjam/settings.py
@@ -48,7 +48,6 @@ INSTALLED_APPS = [
     "peachjam.apps.PeachJamConfig",
     "peachjam_search.apps.PeachjamSearchConfig",
     "peachjam_api.apps.PeachjamApiConfig",
-    "peachjam_ml.apps.PeachjamMLConfig",
     "allauth",
     "allauth.account",
     "allauth.socialaccount",
