+++ peachjam/templates/peachjam/_related_documents.html
@@ -38,3 +38,4 @@
     </div>
   {% endif %}
 </div>
+{% block extra-related-documents %}{% endblock %}
