+++ peachjam/templates/peachjam/layouts/document_detail.html
@@ -107,22 +107,24 @@
                 </button>
               </li>
             {% endif %}
-            {% if n_relationships %}
-              <li class="nav-item" role="presentation">
-                <button class="nav-link"
-                        data-bs-toggle="tab"
-                        data-bs-target="#related-tab"
-                        data-track-event="Document | Tab | Related documents"
-                        type="button"
-                        role="tab"
-                        aria-controls="related-tab"
-                        aria-selected="false">
-                  <i class="bi bi-pj pj-related"></i>
-                  {% trans "Related documents" %}
-                  <span class="badge bg-secondary">{{ n_relationships }}</span>
-                </button>
-              </li>
-            {% endif %}
+            {% block related-documents-tab %}
+              {% if show_related_documents %}
+                <li class="nav-item" role="presentation">
+                  <button class="nav-link"
+                          data-bs-toggle="tab"
+                          data-bs-target="#related-tab"
+                          data-track-event="Document | Tab | Related documents"
+                          type="button"
+                          role="tab"
+                          aria-controls="related-tab"
+                          aria-selected="false">
+                    <i class="bi bi-pj pj-related"></i>
+                    {% trans "Related documents" %}
+                    {% if n_relationships %}<span class="badge bg-secondary">{{ n_relationships }}</span>{% endif %}
+                  </button>
+                </li>
+              {% endif %}
+            {% endblock %}
             {% if cited_documents or documents_citing_current_doc %}
               <li class="nav-item" role="presentation">
                 <button class="nav-link"
@@ -460,14 +462,16 @@
             <div class="container">{% include 'peachjam/_timeline_events.html' %}</div>
           </div>
         {% endif %}
-        {% if n_relationships %}
-          <div class="tab-pane fade"
-               id="related-tab"
-               role="tabpanel"
-               aria-labelledby="related-tab">
-            <div class="container">{% include 'peachjam/_related_documents.html' %}</div>
-          </div>
-        {% endif %}
+        {% block related-documents-tab-pane %}
+          {% if show_related_documents %}
+            <div class="tab-pane fade"
+                 id="related-tab"
+                 role="tabpanel"
+                 aria-labelledby="related-tab">
+              <div class="container">{% include 'peachjam/_related_documents.html' %}</div>
+            </div>
+          {% endif %}
+        {% endblock %}
         {% if cited_documents or documents_citing_current_doc %}
           <div class="tab-pane fade"
                id="citations-tab"
