+++ peachjam/views/documents.py
@@ -1,14 +1,12 @@
 from cobalt import FrbrUri
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
-from django.db.models import F
 from django.http import Http404, HttpResponse
 from django.http.response import FileResponse
 from django.shortcuts import get_list_or_404, get_object_or_404, redirect, reverse
 from django.utils.decorators import method_decorator
 from django.utils.translation import get_language
 from django.views.generic import DetailView, View
-from pgvector.django import MaxInnerProduct
 
 from peachjam.helpers import add_slash, add_slash_to_frbr_uri
 from peachjam.helpers import get_language as get_language_from_request
@@ -21,7 +19,6 @@ from peachjam.models import (
 from peachjam.registry import registry
 from peachjam.resolver import resolver
 from peachjam.views import BaseDocumentDetailView
-from peachjam_ml.models import DocumentEmbedding
 
 
 class DocumentDetailViewResolver(View):
@@ -250,36 +247,6 @@ class DocumentCitationsView(DetailView):
         context["nature"] = nature
         context["direction"] = direction
 
-        return context
-
-
-@method_decorator(add_slash_to_frbr_uri(), name="setup")
-class SimilarDocumentsView(DetailView):
-    template_name = "peachjam/_similar_documents.html"
-    slug_url_kwarg = "frbr_uri"
-    slug_field = "expression_frbr_uri"
-    model = CoreDocument
-
-    def get_similar_documents(self):
-        embedding = get_object_or_404(DocumentEmbedding, document_id=self.object.pk)
-        similar_docs = (
-            DocumentEmbedding.objects.exclude(document_id=self.object.pk)
-            .exclude(text_embedding__isnull=True)
-            .select_related("document")
-            .annotate(
-                similarity=MaxInnerProduct("text_embedding", embedding.text_embedding)
-                * -1,
-                title=F("document__title"),
-                expression_frbr_uri=F("document__expression_frbr_uri"),
-            )
-            .values("title", "expression_frbr_uri", "similarity")
-            .order_by("-similarity")[:10]
-        )
-        return similar_docs
-
-    def get_context_data(self, **kwargs):
-        context = super().get_context_data(**kwargs)
-        context["similar_documents"] = self.get_similar_documents()
         return context
