+++ peachjam_ml/templates/peachjam/_related_documents.html
@@ -1,0 +1,11 @@
+{% extends "peachjam/_related_documents.html" %}
+{% load peachjam %}
+{% block extra-related-documents %}
+  <div class="row">
+    <div class="col-lg">
+      <div hx-get="{% url 'document_similar_docs' document.expression_frbr_uri|strip_first_character %}"
+           hx-trigger="load">
+      </div>
+    </div>
+  </div>
+{% endblock %}
