+++ peachjam_ml/templates/peachjam/layouts/document_detail.html
@@ -1,0 +1,20 @@
+{% extends 'peachjam/layouts/document_detail.html' %}
+{% load i18n %}
+{% block related-documents-tab %}
+  {% if show_related_documents %}
+    <li class="nav-item" role="presentation">
+      <button class="nav-link"
+              data-bs-toggle="tab"
+              data-bs-target="#related-tab"
+              data-track-event="Document | Tab | Related documents"
+              type="button"
+              role="tab"
+              aria-controls="related-tab"
+              aria-selected="false">
+        <i class="bi bi-pj pj-related"></i>
+        {% trans "Related documents" %}
+        {% if n_relationships %}<span class="badge bg-secondary">{{ n_relationships }}</span>{% endif %}
+      </button>
+    </li>
+  {% endif %}
+{% endblock %}
